refactor(sideboard-guide): extract wx-free CSV codec + relocate parser (#806)
Split widgets/frames/app_frame/handlers/sideboard_guide_io.py (343 LOC):

- New sideboard_guide_csv.py: pure wx-free codec with free functions
  export_guide_to_csv(entries, exclusions, zone_cards, file_path) and
  import_guide_from_csv(file_path, mainboard_names, sideboard_names) ->
  (entries, warnings). Mirrors the layer's reads/writes/schema codec
  convention and is directly unit-testable.
- sideboard_guide_io.py (343 -> 166): slim SideboardGuideImportExportHandlers
  mixin keeps only _on_export_guide / _on_import_guide (file dialogs, the
  ad-hoc import-options dialog, status, threaded workers, the apply()
  merge/dedup + persist + refresh logic). Delegates serialization to the
  codec; instance state (sideboard_guide_entries, sideboard_exclusions,
  zone_cards, mainboard/sideboard names) is now passed in as plain args.
- Relocated _parse_card_text into sideboard_guide_persistence.py, its sole
  consumer (three call sites unchanged via self).
- Added tests/test_sideboard_guide_csv.py covering matrix build, exclusions,
  empty-export error, roundtrip, missing-card warnings, bad header, and
  decklist-section termination.

Behavior-preserving: code moved, not rewritten; threading contract preserved
(only the pure codec call runs off-thread, all wx/state mutation stays on the
main thread via wx.CallAfter).
